# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Astral Modulator Public API
  version: 2.0.0-scaffold
  license:
    name: MIT
    identifier: MIT
  description: |
    人类与多 LLM Agent 的协作控制面。CLI（astral-cli 仓库）与 Web GUI 共用本协议。

    v2（容器化任务树，TODO.md D15）：workspace 与 task 都是容器，子任务集合
    统一为 GET/POST <container>/children；全量平铺列表与 /tasks/search 已移除，
    平面查询走 /workspaces/{id}/task-search。protocol_version 升为 2，
    与 v1 不兼容（破坏性变更，开发期不做兼容窗口）。

    已裁决的文档分歧（详见 TODO.md「文档分歧裁决」D1-D4）：
    - Device flow 采用 `/auth/device/authorizations*` 路径。
    - 当前身份端点为 `/auth/me`。

    通用约定：
    - 所有时间戳为 RFC3339 UTC。
    - ID 为 `prefix_uuidv7` 形式的不透明字符串，客户端不得解析内部结构。
    - 错误响应统一使用 ErrorEnvelope；CLI 依赖稳定 error.code。
    - 乐观并发统一使用 body 内 expected_revision（protocol §6 方案 A）。
servers:
  - url: /api/v1
    description: 路径前缀固定 /api/v1（不随协议大版本变化）；协议大版本以
      /.well-known/astral 的 protocol_version 为准
...
security:
  - bearerAuth: []
paths:
...
  /workspaces/{workspace_id}/children:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceId" }
    post:
      tags: [tasks]
      summary: 创建根层任务（投递进 workspace 容器；支持 Idempotency-Key）
      operationId: createRootTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreate" }
      responses:
        "201":
          description: 已创建（revision=1）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404": { $ref: "#/components/responses/Error" }   # tags 含未知名字
    get:
      tags: [tasks]
      summary: |
        workspace 容器的子任务集合 = 根层任务（parent_id 为空），仅此一层。
        子层用 GET /tasks/{task_id}/children 逐容器下钻（D15：不做全量平铺）。
      operationId: listWorkspaceChildren
      parameters:
        - { $ref: "#/components/parameters/TaskStatusFilter" }
        - { $ref: "#/components/parameters/TaskTagFilter" }
        - { $ref: "#/components/parameters/TaskAssigneeFilter" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: 根层任务分页列表（行内含 tags 与 children_count）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskPage"
  /tasks/{task_id}/children:
    parameters:
      - { $ref: "#/components/parameters/TaskId" }
    post:
      tags: [tasks]
      summary: 创建子任务（投递进 task 容器；支持 Idempotency-Key）
      operationId: createChildTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreate" }
      responses:
        "201":
          description: 已创建（parent=容器任务，revision=1）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404": { $ref: "#/components/responses/Error" }   # 容器任务不存在 / tags 含未知名字
    get:
      tags: [tasks]
      summary: task 容器的子任务集合（直接子层，仅此一层；与 workspace children 同参数同响应）
      operationId: listTaskChildren
      parameters:
        - { $ref: "#/components/parameters/TaskStatusFilter" }
        - { $ref: "#/components/parameters/TaskTagFilter" }
        - { $ref: "#/components/parameters/TaskAssigneeFilter" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: 子任务分页列表（行内含 tags 与 children_count）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskPage"
        "404": { $ref: "#/components/responses/Error" }
  /workspaces/{workspace_id}/task-search:
    get:
      tags: [tasks]
      summary: |
        workspace 级平面查询（跨层级）。结构化（tag/status/assignee）与内容
        （regex/fuzzy）平权，至少提供其一（防全量 dump）。语义固定：
        权限过滤 -> 结构化过滤 -> regex 过滤 -> fuzzy 排序 -> 分页（architecture §13）。
        候选集封顶（实现见 TODO.md D7）。
      operationId: searchTasks
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/TaskRegexFilter" }
        - { $ref: "#/components/parameters/TaskFuzzyFilter" }
        - { name: parent_id, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/TaskTagFilter" }
        - { $ref: "#/components/parameters/TaskStatusFilter" }
        - { $ref: "#/components/parameters/TaskAssigneeFilter" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: 查询结果（fuzzy 时 items[].score；行内含 tags 与 children_count）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSearchPage"
        "400": { $ref: "#/components/responses/Error" }   # 零条件 / 非法 regex / 超长输入
  /tasks/{task_id}:
    parameters:
      - { $ref: "#/components/parameters/TaskId" }
    get:
      tags: [tasks]
      summary: 任务详情
      operationId: getTask
      responses:
        "200":
          description: 详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [tasks]
      summary: 更新任务（乐观并发）
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskUpdate" }
      responses:
        "200":
          description: 已更新（revision+1）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "409": { $ref: "#/components/responses/Error" }   # REVISION_CONFLICT，details 带当前 revision
  /tasks/{task_id}/claim:
    post:
      tags: [tasks]
      summary: 原子认领（Phase 1 spike 验收核心）
      operationId: claimTask
      parameters:
        - { $ref: "#/components/parameters/TaskId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_revision]
              properties:
                expected_revision: { type: integer, format: int64, minimum: 1 }
                lease_seconds: { type: integer, minimum: 30, maximum: 3600, default: 300 }
      responses:
        "200":
          description: 认领成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ClaimResult" }
        "409": { $ref: "#/components/responses/Error" }   # TASK_ALREADY_CLAIMED / REVISION_CONFLICT
  /tasks/{task_id}/lease/renew:
    post:
      tags: [tasks]
      summary: 续租（仅 holder）
      operationId: renewLease
      parameters:
        - { $ref: "#/components/parameters/TaskId" }
      responses:
        "200":
          description: 已续租
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lease" }
        "409": { $ref: "#/components/responses/Error" }   # TASK_LEASE_EXPIRED 需重新 claim
  /tasks/{task_id}/lease:
    delete:
      tags: [tasks]
      summary: 释放（holder；human force release 走 override/approval）
      operationId: releaseLease
      parameters:
        - { $ref: "#/components/parameters/TaskId" }
      responses:
        "204": { description: 已释放 }
...
  /tasks/{task_id}/tags/{tag_id}:
    parameters:
      - { name: task_id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
      - { name: tag_id, in: path, required: true, schema: { $ref: "#/components/schemas/Id" } }
    put:
      tags: [tasks]
      summary: 挂载 tag（幂等：已挂返回当前任务，不 bump revision；D11）
      operationId: attachTaskTag
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                expected_revision: { type: integer, format: int64 }
      responses:
        "200":
          description: 已挂载（含 tags 的任务视图）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [tasks]
      summary: 摘除 tag（幂等：未挂返回 204）
      operationId: detachTaskTag
      responses:
        "204": { description: 已摘除或本就未挂载 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Human access token 或 Agent/Service credential（不共用）
    cookieAuth:
      type: apiKey
      in: cookie
      name: astral_session
      description: Web GUI HttpOnly+Secure+SameSite session（与 CLI Bearer 同 actor 不同 session）
  parameters:
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
    TaskId:
      name: task_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, description: 不透明游标，取自 next_cursor }
    TaskStatusFilter:
      name: status
      in: query
      schema: { $ref: "#/components/schemas/TaskStatus" }
    TaskTagFilter:
      name: tag
      in: query
      schema: { type: string }
    TaskAssigneeFilter:
      name: assignee
      in: query
      schema: { type: string }
    TaskRegexFilter:
      name: regex
      in: query
      schema: { type: string, maxLength: 512 }
    TaskFuzzyFilter:
      name: fuzzy
      in: query
      schema: { type: string, maxLength: 256 }
  responses:
    Error:
      description: 统一错误 envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: 次级资源不存在（404 NOT_FOUND，语义见 error.code）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: 状态冲突（409：REVISION_CONFLICT / APPROVAL_EXPIRED /
        TASK_ALREADY_CLAIMED 等，细分语义见 error.code 与 error.details）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    Id:
      type: string
      pattern: "^[a-z]{2,3}_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: 不透明 ID：prefix_uuidv7。前缀 usr/agt/svc/srv/ws/tsk/tgp/tag/doc/cfl/prs/msg/evt/cred/ses/dev/req/aud/apv/inv。
      examples: ["ws_0192ab34-56cd-7ef8-9a01-234567890abc"]
    IdOrNull:
      oneOf:
        - $ref: "#/components/schemas/Id"
        - type: "null"
...
    LeaseOrNull:
      oneOf:
        - $ref: "#/components/schemas/Lease"
        - type: "null"
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable, request_id]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }
            request_id: { type: string }
...
    TaskStatus:
      type: string
      enum: [open, in_progress, blocked, review, done, cancelled]
    TaskPriority:
      type: string
      enum: [low, normal, high, urgent]
    TaskCreate:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: string, default: "" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        tags:
          type: array
          items: { type: string }
          description: |
            已存在的 tag 名（按规范化名解析；v2 中 parent 由容器端点寻址，
            不再有 parent_id 字段）。任一名字不存在 → 404 NOT_FOUND，整体不创建。
    TaskUpdate:
      type: object
      required: [expected_revision]
      properties:
        expected_revision: { type: integer, format: int64, minimum: 1 }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        parent_id: { $ref: "#/components/schemas/IdOrNull" }
        assignee_actor_id: { $ref: "#/components/schemas/IdOrNull" }
    Task:
      type: object
      required: [id, workspace_id, title, description, status, priority, parent_id,
        assignee_actor_id, revision, tags, children_count, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        workspace_id: { $ref: "#/components/schemas/Id" }
        parent_id: { $ref: "#/components/schemas/IdOrNull" }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/TaskPriority" }
        assignee_actor_id: { $ref: "#/components/schemas/IdOrNull" }
        revision: { type: integer, format: int64 }
        tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
          description: v2 起所有 Task 响应恒填充（无 tag 时空数组；修订 D11）
        children_count:
          type: integer
          format: int64
          description: 直接子任务数（UI 展开徽标；恒填充）
        lease: { $ref: "#/components/schemas/LeaseOrNull" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Lease:
      type: object
      required: [holder_actor_id, expires_at]
      properties:
        holder_actor_id: { $ref: "#/components/schemas/Id" }
        expires_at: { type: string, format: date-time }
        renewed_at: { type: string, format: date-time }
    ClaimResult:
      type: object
      required: [task, lease]
      properties:
        task: { $ref: "#/components/schemas/Task" }
        lease: { $ref: "#/components/schemas/Lease" }
    Tag:
      type: object
      required: [id, workspace_id, name]
      properties:
        id: { $ref: "#/components/schemas/Id" }
        workspace_id: { $ref: "#/components/schemas/Id" }
        name: { type: string }
...
    PageMeta:
      type: object
      required: [next_cursor]
      properties:
        next_cursor:
          type: [string, "null"]
          description: null 表示没有更多
...
    TaskPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/Task" } }
    TaskSearchPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                allOf:
                  - $ref: "#/components/schemas/Task"
                  - type: object
                    properties:
                      score: { type: number, description: fuzzy 排序分，仅 fuzzy 查询时出现 }
...